Rename Role enum file for consistency
