+++ src/controllers/Professor.controller.ts
@@ -36,7 +36,7 @@ import {
 import { ProfessorService } from '../services/Professor.service';
 import { ProfessorDto } from '../dto/Professor.dto';
 import { Roles } from '../decorators/Roles.decorator';
-import { Role } from '../enums/Roles.enum';
+import { Role } from '../enums/Role.enum';
 import {
     ApiBearerAuth,
     ApiBody,
