+++ src/entities/User.entity.ts
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 import { Column, Entity, ObjectIdColumn } from 'typeorm';
-import { Role } from '../enums/Roles.enum';
+import { Role } from '../enums/Role.enum';
 
 @Entity()
 export class User {
